Move base conditioner/transformer classes to respective packages.
