+++ tfep/nn/flows/autoregressive.py
@@ -14,106 +14,12 @@ Base autoregressive flow layer for PyTorch.
 # GLOBAL IMPORTS
 # =============================================================================
 
-import abc
 from collections.abc import Sequence
 from typing import Optional
 
 import torch
 
 from tfep.utils.misc import ensure_tensor_sequence
-
-
-# =============================================================================
-# INTERFACES
-# =============================================================================
-
-class Conditioner(torch.nn.Module):
-    """A conditioner for an autoregressive flow.
-
-    This class documents the API of a conditioner layer compatible with an
-    :class:`.AutoregressiveFlow`.
-
-    """
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        """Compute the parameters for the transformer.
-
-        Parameters
-        ----------
-        x : torch.Tensor
-            Shape ``(batch_size, n_features)``. The input features.
-
-        Returns
-        -------
-        parameters : torch.Tensor
-            Shape ``(batch_size, n_parameters)``. The parameters for the transformer.
-
-        """
-        return super().forward(x)  # Raises NotImplementedError.
-
-    @abc.abstractmethod
-    def set_output(self, output: torch.Tensor):
-        """Sets the parameters of the conditioner to produce a constant output.
-
-        This is used to force the autoregressive flow to implement the identity
-        function on initialization.
-
-        Parameters
-        ----------
-        output : torch.Tensor
-            Shape ``(n_parameters,)``. The desired output of the conditioner.
-
-        """
-        pass
-
-
-class Transformer(torch.nn.Module):
-    """A transformer for an autoregressive flow.
-
-    This class documents the API of a transformer layer compatible with an
-    :class:`.AutoregressiveFlow`.
-
-    """
-
-    def forward(self, x: torch.Tensor, parameters: torch.Tensor) -> tuple[torch.Tensor]:
-        """Compute the parameters for the transformer.
-
-        Parameters
-        ----------
-        x : torch.Tensor
-            Shape ``(batch_size, n_features)``. The input features.
-        parameters : torch.Tensor
-            Shape ``(batch_size, n_parameters)``. The parameters for the
-            transformation generated by a conditioner.
-
-        Returns
-        -------
-        y : torch.Tensor
-            Shape ``(batch_size, n_features)``. The transformed features.
-        log_det_J : torch.Tensor
-            Shape ``(batch_size,)``. The log absolute value of the Jacobian
-            determinant of the transformation.
-
-        """
-        return super().forward(x)  # Raises NotImplementedError.
-
-    @abc.abstractmethod
-    def get_parameters_identity(self, n_features: int) -> torch.Tensor:
-        """Returns the parameters that would make this transformer the identity function.
-
-        Parameters
-        ----------
-        n_features : int
-            The number of transformed features.
-
-        Returns
-        -------
-        parameters_identity : torch.Tensor
-            Shape ``(n_parameters,)``. The parameters of the transformation that
-            makes this transformer the identity function.
-
-        """
-        pass
 
 
 # =============================================================================
